アップロードページにフッターを表示できるようにした。
